Ares Refactored Routes and Rank Names

+++ src/Article/Service/DeleteCommentService.php
@@ -13,6 +13,9 @@ use Ares\Article\Repository\CommentRepository;
 use Ares\Framework\Exception\DataObjectManagerException;
 use Ares\Framework\Interfaces\CustomResponseInterface;
 use Ares\Framework\Interfaces\HttpResponseCodeInterface;
+use Ares\User\Repository\UserRepository;
+use Ares\Article\Entity\Comment;
+use Ares\User\Entity\User;
 
 /**
  * Class DeleteArticleService
@@ -27,7 +30,8 @@ class DeleteCommentService
      * @param CommentRepository $commentRepository
      */
     public function __construct(
-        private CommentRepository $commentRepository
+        private CommentRepository $commentRepository,
+        private UserRepository $userRepository
     ) {}
 
     /**
@@ -37,19 +41,28 @@ class DeleteCommentService
      * @throws CommentException
      * @throws DataObjectManagerException
      */
-    public function execute(int $id): CustomResponseInterface
+    public function execute(int $id, int $userId): CustomResponseInterface
     {
-        $deleted = $this->commentRepository->delete($id);
+        /** @var User $user */
+        $user = $this->userRepository->get($userId);
+        
+        /** @var Comment $comment */
+        $comment = $this->commentRepository->get($userId);
 
-        if (!$deleted) {
-            throw new CommentException(
-                __('Comment could not be deleted'),
-                ArticleResponseCodeInterface::RESPONSE_ARTICLE_COMMENT_NOT_DELETED,
-                HttpResponseCodeInterface::HTTP_RESPONSE_UNPROCESSABLE_ENTITY
-            );
+        if($comment->getUserId() == $userId || $user->hasPermission('manage-comments')) {
+            $deleted = $this->commentRepository->delete($id);
+
+            if (!$deleted) {
+                throw new CommentException(
+                    __('Comment could not be deleted'),
+                    ArticleResponseCodeInterface::RESPONSE_ARTICLE_COMMENT_NOT_DELETED,
+                    HttpResponseCodeInterface::HTTP_RESPONSE_UNPROCESSABLE_ENTITY
+                );
+            }
+
+            return response()->setData(true);
         }
 
-        return response()
-            ->setData(true);
+        return response()->setData(false);
     }
 }
